requirements.txt
hydra-core
matplotlib
mne
mne_connectivity
numpy
omegaconf
pandas
pytorch_lightning
PyYAML
PyYAML
scipy
seaborn
six
torch
wandb
nibabel